+++ src/script_error.rs
@@ -1,17 +1,6 @@
-use secp256k1;
 use thiserror::Error;
 
-/// Things that can go wrong when constructing a `HashType` from bit flags.
-#[derive(Copy, Clone, PartialEq, Eq, Debug, Error)]
-pub enum InvalidHashType {
-    /// Either or both of the two least-significant bits must be set.
-    #[error("unknowned signed outputs")]
-    UnknownSignedOutputs,
-    /// With v5 transactions, bits other than those specified for `HashType` must be 0. The `i32`
-    /// includes only the bits that are undefined by `HashType`.
-    #[error("extra bits set")]
-    ExtraBitsSet(i32),
-}
+use crate::signature;
 
 #[derive(Copy, Clone, PartialEq, Eq, Debug, Error)]
 pub enum ScriptNumError {
@@ -23,10 +12,9 @@ pub enum ScriptNumError {
 }
 
 #[derive(Copy, Clone, PartialEq, Eq, Debug, Error)]
-#[repr(i32)]
 pub enum ScriptError {
     #[error("Ok")]
-    Ok = 0, // Unused (except in converting the C++ error to Rust)
+    Ok, // Unused (except in converting the C++ error to Rust)
 
     #[error("unknown error")]
     UnknownError,
@@ -95,27 +83,14 @@ pub enum ScriptError {
     #[error("unsatisfied locktime condition")]
     UnsatisfiedLockTime,
 
-    // BIP62
-    #[error(
-        "{}",
-        .0.map_or(
-            "unknown signature hash type error".to_owned(),
-            |iht| format!("signature hash type error: {}", iht)
-        )
-    )]
-    SigHashType(Option<InvalidHashType>),
-
-    #[error("signature DER encoding error")]
-    SigDER(Option<secp256k1::Error>),
+    #[error("signature encoding error: {}", .0)]
+    SignatureEncoding(signature::Error),
 
     #[error("minimal data requirement not met")]
     MinimalData,
 
     #[error("signature push only requirement not met")]
     SigPushOnly,
-
-    #[error("signature s value is too high")]
-    SigHighS,
 
     #[error("signature null dummy error")]
     SigNullDummy,
@@ -147,5 +122,11 @@ pub enum ScriptError {
 impl From<ScriptNumError> for ScriptError {
     fn from(value: ScriptNumError) -> Self {
         ScriptError::ScriptNumError(value)
+    }
+}
+
+impl From<signature::Error> for ScriptError {
+    fn from(value: signature::Error) -> Self {
+        ScriptError::SignatureEncoding(value)
     }
 }